writer: always escape \n and \r inside quoted strings
so all quoted strings are on the same line.
glyphsLib writer does that. Also json does the same.

I still pass the literal \t (tab) character unescaped. I think it doesn't
do any harm, though it is a bit odd since we are escaping all the rest of the
control chars...