write strings unquoted when not a number even if start with digit
the get_unquoted_string_type function from the parser now checks all the characters
and returns whether the string contains an int, float or neither.
The same function is used in the writer to decide whether to write a string with
quotes when it appears like a number.
However, if the string is actually not a number (e.g. 7z) but contains only
valid unquoted characters, it can be left unquoted.